+++ bpdm-common-test/src/main/kotlin/org/eclipse/tractusx/bpdm/test/testdata/gate/BusinessPartnerVerboseValues.kt
@@ -234,13 +234,13 @@ object BusinessPartnerVerboseValues {
         deliveryServiceType = DeliveryServiceType.PO_BOX,
         deliveryServiceQualifier = "DHL",
         deliveryServiceNumber = "1234",
-        geographicCoordinates = GeoCoordinateDto(7.619f, 45.976f, 4478f),
+        geographicCoordinates = GeoCoordinateDto(7.619, 45.976, 4478.0),
         postalCode = "70547",
         administrativeAreaLevel1 = "adminAreaLevel1RegionCode_2",
     )
 
     val postalAddress1 = PhysicalPostalAddressDto(
-        geographicCoordinates = GeoCoordinateDto(13.178f, 48.946f),
+        geographicCoordinates = GeoCoordinateDto(13.178, 48.946),
         country = CountryCode.DE,
         postalCode = "70546 ",
         city = "Stuttgart",
@@ -257,7 +257,7 @@ object BusinessPartnerVerboseValues {
     )
 
     val postalAddress2 = PhysicalPostalAddressDto(
-        geographicCoordinates = GeoCoordinateDto(7.619f, 45.976f, 4478f),
+        geographicCoordinates = GeoCoordinateDto(7.619, 45.976, 4478.0),
         country = CountryCode.US,
         postalCode = "70547",
         city = "Atlanta",
@@ -274,7 +274,7 @@ object BusinessPartnerVerboseValues {
     )
 
     val postalAddress3 = PhysicalPostalAddressDto(
-        geographicCoordinates = GeoCoordinateDto(13.178f, 48.946f),
+        geographicCoordinates = GeoCoordinateDto(13.178, 48.946),
         country = CountryCode.DE,
         postalCode = "70546 ",
         city = "Stuttgart",
@@ -337,7 +337,7 @@ object BusinessPartnerVerboseValues {
 
     //New Values for Logistic Addresses Tests
     val postalAddressLogisticAddress1 = PhysicalPostalAddressDto(
-        geographicCoordinates = GeoCoordinateDto(13.178f, 48.946f),
+        geographicCoordinates = GeoCoordinateDto(13.178, 48.946),
         country = CountryCode.DE,
         postalCode = "70546 ",
         city = "Stuttgart",
@@ -354,7 +354,7 @@ object BusinessPartnerVerboseValues {
     )
 
     val postalAddressLogisticAddress2 = PhysicalPostalAddressDto(
-        geographicCoordinates = GeoCoordinateDto(7.619f, 45.976f, 4478f),
+        geographicCoordinates = GeoCoordinateDto(7.619, 45.976, 4478.0),
         country = CountryCode.US,
         postalCode = "70547",
         city = "Atlanta",
@@ -752,7 +752,7 @@ object BusinessPartnerVerboseValues {
             name = "Address Name",
             addressType = AddressType.AdditionalAddress,
             physicalPostalAddress = PhysicalPostalAddressDto(
-                geographicCoordinates = GeoCoordinateDto(0.5f, 0.5f, 0.5f),
+                geographicCoordinates = GeoCoordinateDto(0.5, 0.5, 0.5),
                 country = CountryCode.PT,
                 administrativeAreaLevel1 = "PT-PT",
                 administrativeAreaLevel2 = "pt-admin-level-2-cleaned",
@@ -779,7 +779,7 @@ object BusinessPartnerVerboseValues {
                 door = "door"
             ),
             alternativePostalAddress = AlternativePostalAddressDto(
-                geographicCoordinates = GeoCoordinateDto(0.6f, 0.6f, 0.6f),
+                geographicCoordinates = GeoCoordinateDto(0.6, 0.6, 0.6),
                 country = CountryCode.PT,
                 administrativeAreaLevel1 = "PT-PT",
                 postalCode = "postal-code-cleaned",