+++ bpdm-orchestrator/src/main/resources/db/migration/V6_2_0_2__update_geographical_coordinates_type.sql
@@ -1,0 +1,17 @@
+ALTER TABLE business_partner_addresses
+ALTER COLUMN alt_latitude SET DATA TYPE double precision;
+
+ALTER TABLE business_partner_addresses
+ALTER COLUMN alt_longitude SET DATA TYPE double precision;
+
+ALTER TABLE business_partner_addresses
+ALTER COLUMN alt_altitude SET DATA TYPE double precision;
+
+ALTER TABLE business_partner_addresses
+ALTER COLUMN phy_latitude SET DATA TYPE double precision;
+
+ALTER TABLE business_partner_addresses
+ALTER COLUMN phy_longitude SET DATA TYPE double precision;
+
+ALTER TABLE business_partner_addresses
+ALTER COLUMN phy_altitude SET DATA TYPE double precision;